Extract legend/linedraw to utils. Refactor
