+++ src/features/CountryDistance/CountryDistance.jsx
@@ -1,8 +1,11 @@
 import React, { useState, useCallback, useEffect } from "react";
 import ReactDom from "react-dom";
 import { parseJSON } from "./parseMapJSON";
-import { LineDraw } from "./lineDraw";
+import { LineDraw } from "../../utils/lineDraw";
+import colorScheme from "../../utils/colorScheme";
+import { get_keys } from "../../model/dataHandler";
 import { distance } from "../../utils/categories";
+import { interpolateRgb } from "d3";
 import { Form, InputGroup, Button } from "react-bootstrap";
 import { useRef } from "react";
 import InfoPopover from "./InfoPopover";
@@ -12,6 +15,32 @@ import InfoPopover from "./InfoPopover";
 
 const canvasWidth = "100%";
 const canvasHeight = "100%";
+
+function countryToColor(country, selected) {
+  if (!country.hasData)
+    return colorScheme.noData;
+  if(!selected) {
+    return colorScheme.middle;
+  }
+
+  var relative_value;
+  var dist_sq = 0;
+  // Choose higher values to make the dimension with the largest distance play a larger role.
+  // Inspired by Shephard Interpolation: https://en.wikipedia.org/wiki/Inverse_distance_weighting#/media/File:Shepard_interpolation_2.png
+  // Image that three countries have the results A: (0,0), B: (0.5, 0.5), C: (0.9)
+  // With exponent = 1, C is closer to A than B as 0.9 < 0.5 + 0.5
+  // With exponent = 2, B is closeras (0.25 + 0.25)^0.5 < 0.9^2^0.5 = 0.9
+  const exponent = 2;
+  for(const k of get_keys()) {
+    dist_sq += Math.abs(country[k] - selected[k]) ** exponent;
+  }
+  relative_value = exponent * dist_sq ** (1/exponent);
+  const extreme_color = relative_value < 0 ? colorScheme.left : colorScheme.right;
+  //between 0 and 1. 0 is white (=similar to selected), 1 is extreme_color (=not similar to selected)
+  const absolute_value = Math.abs(relative_value);
+  return interpolateRgb(colorScheme.middle, extreme_color)(absolute_value);
+};
+
 
 export default function CountryDistance({activeTab}) {
   //currently selected country
@@ -43,8 +72,8 @@ export default function CountryDistance({activeTab}) {
 
   const svg = (
       <svg width={canvasWidth} height={canvasHeight} ref={svgRef} onMouseLeave={() => { setHovered(null) } }>
+          {svgHasMounted &&
           <>
-              {svgHasMounted &&
               <LineDraw
                 data={mapData}
                 selectCountry={setSelected}
@@ -56,9 +85,10 @@ export default function CountryDistance({activeTab}) {
                 zoomLevelSetter={zoomLevelSetter}
                 doReset={doReset}
                 setDoReset={setDoReset}
+                countryToColor={countryToColor}
               />
-              }
           </>
+          }
       </svg>
   );
   return (
