+++ src/utils/colorScheme.jsx
@@ -1,0 +1,8 @@
+export default {
+  left: '#ef4400',
+  middle: '#f7f7f7',
+  right: '#0083cf',
+  selectedCountry: '#00A600',
+  hoveredCountry: '#00CC00',
+  noData: 'gray',
+};
